backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.10.4
sentence-transformers==2.7.0
chromadb==0.4.18
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
huggingface-hub>=0.16.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
